refactor: submission form schema, remove unused files
