+++ package.json
@@ -19,11 +19,11 @@
     "@chakra-ui/react": "^2.9.4",
     "@emotion/react": "^11.11.3",
     "@emotion/styled": "^11.11.0",
-    "@hookform/resolvers": "^3.9.0",
+    "@hookform/resolvers": "^3.9.1",
     "@next-auth/prisma-adapter": "^1.0.7",
     "@prisma/client": "5.21.1",
     "@sentry/nextjs": "^8.35.0",
-    "@solana/spl-token": "^0.4.8",
+    "@solana/spl-token": "^0.4.9",
     "@solana/wallet-adapter-base": "^0.9.23",
     "@solana/wallet-adapter-react": "^0.15.35",
     "@solana/wallet-adapter-react-ui": "^0.9.35",
@@ -58,6 +58,7 @@
     "immer": "^10.1.1",
     "jotai": "^2.10.1",
     "jsonwebtoken": "^9.0.2",
+    "knip": "^5.36.3",
     "lodash.debounce": "^4.0.8",
     "micro": "^10.0.1",
     "next": "14.2.14",
@@ -77,7 +78,7 @@
     "react-countdown": "^2.3.5",
     "react-dom": "18.3.1",
     "react-dropzone": "^14.2.10",
-    "react-hook-form": "^7.53.1",
+    "react-hook-form": "^7.52.2",
     "react-icons": "^5.2.1",
     "react-intersection-observer": "^9.13.0",
     "react-select": "^5.8.0",
@@ -86,7 +87,7 @@
     "sass": "^1.80.4",
     "sharp": "^0.32.6",
     "slugify": "^1.6.6",
-    "sonner": "^1.5.0",
+    "sonner": "^1.7.0",
     "svix": "^1.37.0",
     "textarea-caret": "^3.1.0",
     "tiptap-extension-resize-image": "^1.1.7",
@@ -94,7 +95,7 @@
     "unfurl.js": "^6.4.0",
     "unique-names-generator": "^4.7.1",
     "zod": "^3.23.8",
-    "zustand": "^5.0.0"
+    "zustand": "^5.0.1"
   },
   "devDependencies": {
     "@tanstack/eslint-plugin-query": "^5.59.7",
