+++ src/features/listings/queries/index.ts
@@ -3,5 +3,4 @@ export * from './recent-earners';
 export * from './region-listings';
 export * from './related-listing';
 export * from './submission-count';
-export * from './submission-details';
 export * from './submissions';
