--- src/features/listings/queries/submission-details.ts
@@ -1,22 +1,0 @@
-import { queryOptions } from '@tanstack/react-query';
-import axios from 'axios';
-
-import { type SubmissionWithUser } from '@/interface/submission';
-
-interface SubmissionDetailsParams {
-  submissionId?: string;
-}
-
-const fetchSubmissionDetails = async (
-  params: SubmissionDetailsParams,
-): Promise<SubmissionWithUser> => {
-  const { data } = await axios.get(`/api/submission/details/`, { params });
-  return data;
-};
-
-export const submissionDetailsQuery = (params: SubmissionDetailsParams) =>
-  queryOptions({
-    queryKey: ['submissionDetails', params],
-    queryFn: () => fetchSubmissionDetails(params),
-    enabled: !!params.submissionId,
-  });
