+++ src/features/listings/utils/status.ts
@@ -1,0 +1,102 @@
+import {
+  type Listing,
+  type ListingWithSubmissions,
+  type StatusFilter,
+} from '../types';
+import { isDeadlineOver } from './deadline';
+
+export function getStatusFilterQuery(statusFilter: StatusFilter | undefined) {
+  let statusFilterQuery = {};
+
+  if (statusFilter) {
+    if (statusFilter === 'open') {
+      statusFilterQuery = {
+        deadline: {
+          gte: new Date(),
+        },
+      };
+    } else if (statusFilter === 'review') {
+      statusFilterQuery = {
+        deadline: {
+          lte: new Date(),
+        },
+        isWinnersAnnounced: false,
+      };
+    } else if (statusFilter === 'completed') {
+      statusFilterQuery = {
+        isWinnersAnnounced: true,
+      };
+    }
+  }
+
+  return statusFilterQuery;
+}
+
+export const getListingDraftStatus = (
+  status: string | undefined,
+  isPublished: boolean | undefined,
+) => {
+  if (status === 'CLOSED') return 'CLOSED';
+  if (status === 'REVIEW') return 'REVIEW';
+  if (status === 'PREVIEW') return 'PREVIEW';
+  if (status === 'VERIFYING') return 'VERIFYING';
+  if (status === 'VERIFY_FAIL') return 'VERIFY_FAIL';
+  if (isPublished) return 'PUBLISHED';
+  return 'DRAFT';
+};
+
+export const getListingTypeLabel = (type: string) => {
+  if (type === 'project') return 'Project';
+  if (type === 'hackathon') return 'Hackathon';
+  if (type === 'bounty') return 'Bounty';
+  if (type === 'grant') return 'Grant';
+  return;
+};
+
+export const getListingStatus = (
+  listing: Listing | ListingWithSubmissions | any,
+  isGrant?: boolean,
+) => {
+  if (!listing) return 'Draft';
+
+  const listingStatus = getListingDraftStatus(
+    listing?.status,
+    listing?.isPublished,
+  );
+  const hasDeadlinePassed = isDeadlineOver(listing?.deadline || '');
+
+  if (listingStatus === 'PREVIEW') return 'Draft';
+  if (listingStatus === 'VERIFYING') return 'Under Verification';
+  if (listingStatus === 'VERIFY_FAIL') return 'Verification Failed';
+  if (listingStatus === 'DRAFT') return 'Draft';
+  if (listing?.type === 'grant' || isGrant) return 'Ongoing';
+
+  switch (listingStatus) {
+    case 'CLOSED':
+      return 'Closed';
+    case 'REVIEW':
+    case 'PUBLISHED':
+      if (!hasDeadlinePassed && !listing?.isWinnersAnnounced)
+        return 'In Progress';
+      if (!listing?.isWinnersAnnounced) return 'In Review';
+      if (
+        listing?.isWinnersAnnounced &&
+        listing?.totalPaymentsMade !== listing?.totalWinnersSelected &&
+        listing?.isFndnPaying
+      )
+        return 'Fndn to Pay';
+      if (
+        listing?.isWinnersAnnounced &&
+        listing?.totalPaymentsMade !== listing?.totalWinnersSelected
+      )
+        return 'Payment Pending';
+      if (
+        listing?.isWinnersAnnounced &&
+        listing?.totalPaymentsMade === listing?.totalWinnersSelected
+      )
+        return 'Completed';
+      return 'In Review';
+    default:
+      return 'Draft';
+  }
+};
