+++ src/features/listings/utils/tweetTemplate.ts
@@ -1,0 +1,6 @@
+export function tweetTemplate(url: string) {
+  return `The results of this latest @SuperteamEarn listing are out. Congratulations to the winners👏
+
+${url}
+`;
+}
