+++ src/pages/api/submission/create.ts
@@ -2,58 +2,97 @@ import type { NextApiResponse } from 'next';
 
 import { type NextApiRequestWithUser, withAuth } from '@/features/auth';
 import { sendEmailNotification } from '@/features/emails';
+import { submissionSchema } from '@/features/listings';
 import logger from '@/lib/logger';
 import { prisma } from '@/prisma';
 import { safeStringify } from '@/utils/safeStringify';
 
 async function submission(req: NextApiRequestWithUser, res: NextApiResponse) {
   const userId = req.userId;
-  const { listingId, link, tweet, otherInfo, eligibilityAnswers, ask } =
-    req.body;
+  const {
+    listingId,
+    applicationLink,
+    tweetLink,
+    otherInfo,
+    eligibilityAnswers,
+    ask,
+  } = req.body;
   logger.debug(`Request body: ${safeStringify(req.body)}`);
   logger.debug(`User: ${safeStringify(req.userId)}`);
 
   try {
-    const existingSubmission = await prisma.submission.findFirst({
-      where: {
-        userId,
-        listingId,
-      },
+    const listing = await prisma.bounties.findUnique({
+      where: { id: listingId },
     });
 
-    if (existingSubmission) {
-      return res.status(400).json({
-        message: 'Submission already exists for this user and listing.',
+    if (!listing) {
+      return res.status(404).json({
+        message: 'Listing not found',
       });
     }
 
-    const result = await prisma.submission.create({
-      data: {
-        userId: userId as string,
-        listingId,
-        link: link || '',
-        tweet: tweet || '',
-        otherInfo: otherInfo || '',
-        eligibilityAnswers: eligibilityAnswers || undefined,
-        ask: ask || null,
-      },
-      include: {
-        listing: {
-          select: {
-            pocId: true,
+    try {
+      const validatedData = submissionSchema(
+        listing as any,
+        listing.minRewardAsk || 0,
+        listing.maxRewardAsk || 0,
+      ).parse({
+        applicationLink,
+        tweetLink,
+        otherInfo,
+        ask,
+        eligibilityAnswers,
+      });
+
+      const existingSubmission = await prisma.submission.findFirst({
+        where: {
+          userId,
+          listingId,
+        },
+      });
+
+      if (existingSubmission) {
+        return res.status(400).json({
+          message: 'Submission already exists for this user and listing.',
+        });
+      }
+
+      const result = await prisma.submission.create({
+        data: {
+          userId: userId as string,
+          listingId,
+          link: validatedData.applicationLink || '',
+          tweet: validatedData.tweetLink || '',
+          otherInfo: validatedData.otherInfo || '',
+          eligibilityAnswers: validatedData.eligibilityAnswers || undefined,
+          ask: validatedData.ask || null,
+        },
+        include: {
+          listing: {
+            select: {
+              pocId: true,
+            },
           },
         },
-      },
-    });
+      });
 
-    sendEmailNotification({
-      type: 'submissionTalent',
-      id: listingId,
-      userId: userId as string,
-      triggeredBy: userId,
-    });
+      sendEmailNotification({
+        type: 'submissionTalent',
+        id: listingId,
+        userId: userId as string,
+        triggeredBy: userId,
+      });
 
-    return res.status(200).json(result);
+      return res.status(200).json(result);
+    } catch (validationError: any) {
+      if (validationError.errors) {
+        return res.status(400).json({
+          message: 'Validation error',
+          errors: validationError.errors,
+        });
+      }
+      throw validationError;
+    }
   } catch (error: any) {
     logger.error(`User ${userId} unable to submit: ${safeStringify(error)}`);
     return res.status(400).json({
